fix: improve name arg in check_scalar_or_sequence
Before, name could be None and was set to name = name or "sequence" and
name = name or "scalar". The latter option was chosen if users passed
name=0 or name=0.0. Now, we choose name: str = "scalar_or_sequence" in
the signature so that the everything is just like in check_scalar and
check_sequence.