fix: resort errors in .utils._checks._check_scalar
In cases where scalar did not match type_ (_CheckError) and was not
compatible with ge, gt, le or lt (TypeError), the former error wasn't
raised but only the latter. Now, the _CheckError comes first as it
should be. The TypeError is only for development purposes. It is meant
to help setting up bulletproof checks. In the end, it should never occur
when using the public API.